requirements.txt
Babel==2.8.0
Click==7.0
Flask==1.1.1
Flask-Babel==0.12.2
httplib2==0.15.0
itsdangerous==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
markdown-katex==201912.11b0
MarkupSafe==1.1.1
pathlib2==2.3.5
plantuml==0.3.0
plantuml-markdown==3.2.1
python-dotenv==0.10.4
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0
